// src/types/resilienceTypes.ts
export interface ResilienceItem {
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  subTitle?: any;
  id: number;
  sectionId?: string;
  title: string;
  content: string;
  enterLabel: string;
  enterHref: string;
  mainImageSrc: string;
  mainImageAlt: string;
  thumbImageSrc: string;
  thumbImageAlt: string;
}